+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- torch algo: test and document the fact that `with_batch_norm_parameters` is only about the running mean and variance of the batch norm layers (#30)
+- torch algo: test that `with_batch_norm_parameters` is only about the running mean and variance of the batch norm layers (#30)
 
 ## [0.31.0](https://github.com/Substra/substrafl/releases/tag/0.31.0) - 2022-10-03